Cleaned up imports.
Fixed torrent validation.

Fixed max piece size (needs to be in bytes, torf when called in script does not do the conversion to MiB for you).

Fixed cli_ui not working and thus breaking the script. CustomOutput needed isatty() method so cli_ui can detect that it's an interactive terminal.